+++ test/fixtures/cookbooks/jenkins_authentication/recipes/credential_file.rb
@@ -37,7 +37,6 @@ file cred_file do
 end
 node.default[:jenkins][:executor][:cli_credential_file] = cred_file
 
-
 jenkins_plugin 'pam-auth' do
   notifies :restart, 'service[jenkins]', :immediately
 end