+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This file is used to list changes made in each version of the jenkins cookbook.
 ## Unreleased
 
 - Include yum-epel cookbook on RHEL platforms for new daemonize package dependency
+- Add new `update_center_sleep` attribute to set the time to wait for updates to quiesce in Jenkins
 
 ## 9.1.0 - *2021-08-11*